feat(nitify): add create task action / new task trigger
