+++ packages/backend/.env
@@ -6,4 +6,4 @@ AP_ENVIRONMENT="dev"
 AP_TRIGGER_DEFAULT_POLL_INTERVAL=1
 AP_TELEMETRY_ENABLED=false
 
-AP_DEV_PIECES="discord,delay,approval,hackernews,github,slack,store,wordpress,tidycal"
+AP_DEV_PIECES="discord,delay,approval,hackernews,github,slack,store,wordpress"
