+++ packages/pieces/nifty/README.md
@@ -1,7 +1,7 @@
-# pieces-nitfy
+# pieces-nifty
 
 This library was generated with [Nx](https://nx.dev).
 
 ## Building
 
-Run `nx build pieces-nitfy` to build the library.
+Run `nx build pieces-nifty` to build the library.
