+++ packages/pieces/nifty/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@activepieces/piece-nifty",
+  "version": "0.0.1"
+}
