+++ packages/pieces/nifty/project.json
@@ -1,0 +1,32 @@
+{
+  "name": "pieces-nifty",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "packages/pieces/nifty/src",
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "@nx/js:tsc",
+      "outputs": ["{options.outputPath}"],
+      "options": {
+        "outputPath": "dist/packages/pieces/nifty",
+        "tsConfig": "packages/pieces/nifty/tsconfig.lib.json",
+        "packageJson": "packages/pieces/nifty/package.json",
+        "main": "packages/pieces/nifty/src/index.ts",
+        "assets": ["packages/pieces/nifty/*.md"],
+        "buildableProjectDepsInPackageJsonType": "dependencies"
+      }
+    },
+    "publish": {
+      "command": "node tools/scripts/publish.mjs piece-nitfy {args.ver} {args.tag}",
+      "dependsOn": ["build"]
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": ["packages/pieces/nifty/**/*.ts"]
+      }
+    }
+  },
+  "tags": []
+}
