--- packages/pieces/nitfy/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@activepieces/piece-nitfy",
-  "version": "0.0.1"
-}
