--- packages/pieces/nitfy/project.json
@@ -1,42 +1,0 @@
-{
-  "name": "pieces-nitfy",
-  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/pieces/nitfy/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": [
-        "{options.outputPath}"
-      ],
-      "options": {
-        "outputPath": "dist/packages/pieces/nitfy",
-        "tsConfig": "packages/pieces/nitfy/tsconfig.lib.json",
-        "packageJson": "packages/pieces/nitfy/package.json",
-        "main": "packages/pieces/nitfy/src/index.ts",
-        "assets": [
-          "packages/pieces/nitfy/*.md"
-        ],
-        "buildableProjectDepsInPackageJsonType": "dependencies"
-      }
-    },
-    "publish": {
-      "command": "node tools/scripts/publish.mjs pieces-nitfy {args.ver} {args.tag}",
-      "dependsOn": [
-        "build"
-      ]
-    },
-    "lint": {
-      "executor": "@nx/linter:eslint",
-      "outputs": [
-        "{options.outputFile}"
-      ],
-      "options": {
-        "lintFilePatterns": [
-          "packages/pieces/nitfy/**/*.ts"
-        ]
-      }
-    }
-  },
-  "tags": []
-}
